client/requirements.txt
boto3==1.26.142
botocore==1.29.142
gTTS==2.3.2
playwright==1.34.0
praw==7.7.0
prawcore~=2.3.0
requests==2.31.0
rich==13.3.5
toml==0.10.2
translators==5.7.6
pyttsx3==2.90
Pillow==9.5.0
tomlkit==0.11.8
Flask==2.3.2
clean-text==0.6.0
unidecode==1.3.6
spacy==3.5.3
torch==2.0.1
transformers==4.29.2
ffmpeg-python==0.2.0
elevenlabs==0.2.16
yt-dlp==2023.3.4
edge-tts==6.1.5
stable-ts==2.6.4
ass== 0.5.3